+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 # SEOmatic Changelog
 
-## 3.3.44 - UNRELEASED
+## 3.3.46 - UNRELEASED
+### Added
+* Added support for pulling from Preparse fields for SEO fields (https://github.com/nystudio107/craft-seomatic/issues/927)
+
+### Fixed
+* Fix an issue where `Undefined class constant ‘EVENT_REGISTER_GQL_SCHEMA_COMPONENTS` would be thrown if you were running Craft CMS < `^3.5.0` (https://github.com/nystudio107/craft-seomatic/issues/933)
+* Empty out the entity image settings to ensure the image gets removed if it no longer exists (https://github.com/nystudio107/craft-seomatic/issues/760)
+
+## 3.3.45 - 2021.07.20
+### Fixed
+* Fixed an issue where `MedicalBusiness` JSON-LD type didn't inherit all of the properties from `MedicalOrganization` that it should
+* Handle per-entry type settings for the sitemap generation (https://github.com/nystudio107/craft-seomatic/issues/929)
+
+## 3.3.44 - 2021.07.09
+### Added
+* Added `truncateTitleTags` & `truncateDescriptionTags` settings to control whether the max title / description lengths should be enforced or not (on by default) (https://github.com/nystudio107/craft-seomatic/issues/922)
+
 ### Changed
 * Changed the Google & Bing sitemap submission URLs to `https` (https://github.com/nystudio107/craft-seomatic/issues/903)
+
+### Fixed
+* Handle the case where `additionalSitemaps` isn't a zero-based index, due to items being deleted from it (https://github.com/nystudio107/craft-seomatic/issues/921)
 
 ## 3.3.43 - 2021.06.08
 ### Added